Harvestable Zones (#18)
Replanting zones now function a bit like growing zones in that pawns will try to keep them clear and plants can be harvested automatically. This last option is largely useless in vanilla, but with Extract Any Plant or similar mods it makes a berry bush "growing zone" possible.

I can't believe how much functionality is hardcoded to the Zone_Growing type. This should not have been as complicated as it was.